Change project struture
