+++ escout/Controllers/GameObjects/ClubController.cs
@@ -9,11 +9,11 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace escout.Controllers
+namespace escout.Controllers.GameObjects
 {
     [Authorize]
     [ApiController]
-    [Route("api/v1")]
+    [Route("api/v1/game-object")]
     public class ClubController : ControllerBase
     {
         private readonly DataContext context;
@@ -96,68 +96,6 @@ namespace escout.Controllers
             {
                 return new NotFoundResult();
             }
-        }
-
-        [HttpGet]
-        [Route("clubStatistics")]
-        [ProducesResponseType(StatusCodes.Status200OK)]
-        public ActionResult<Statistics> GetAthleteStatistics(int clubId, int? gameId)
-        {
-            var gameEvents = new List<GameEvent>();
-            var totalStatistics = new Statistics();
-            var totalEvents = new List<GameEvent>();
-            var count = new List<Counter>();
-
-            if (gameId != null)
-                gameEvents = context.gameEvents.Where(x => x.clubId == clubId && x.gameId == gameId).ToList();
-            else
-                gameEvents = context.gameEvents.Where(x => x.clubId == clubId).ToList();
-
-            var uniqueGames = gameEvents.Select(x => x.gameId).Distinct();
-
-            foreach (int i in uniqueGames)
-            {
-                var game = gameEvents.Where(x => x.gameId == i).ToList();
-                foreach (Event e in context.events)
-                {
-                    var events = game.Where(x => x.eventId == e.id).ToList();
-                    var gameStats = new GameStats
-                    {
-                        EventId = e.id,
-                        Count = events.Count(),
-                        GameId = i
-                    };
-
-                    var counter = new Counter
-                    {
-                        Count = gameStats.Count,
-                        EventId = e.id
-                    };
-
-                    count.Add(counter);
-                    totalStatistics.GameStats.Add(gameStats);
-                }
-            }
-
-            foreach (Event e in context.events)
-            {
-                var counter = new List<Counter>();
-                counter = count.Where(x => x.EventId == e.id).ToList();
-                totalEvents = gameEvents.Where(x => x.eventId == e.id).ToList();
-
-                var totalStats = new TotalStats
-                {
-                    EventId = e.id,
-                    Count = totalEvents.Count(),
-                    Average = counter.Average(x => x.Count),
-                    StandardDeviation = GameStatistics.StdDev(counter.Select(x => x.Count).ToArray()),
-                    Median = GameStatistics.Median(counter.Select(x => x.Count).ToArray())
-                };
-
-                totalStatistics.TotalStats.Add(totalStats);
-            }
-
-            return totalStatistics;
         }
     }
 }
