+++ escout/Helpers/Configurations.cs
