+++ escoutTests/Controllers/FavoritesControllerTests.cs
@@ -1,4 +1,5 @@
-﻿using escout.Models;
+﻿using escout.Controllers.GameObjects;
+using escout.Models;
 using escoutTests.Resources;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
