+++ escoutTests/Controllers/UserControllerTests.cs
@@ -1,4 +1,5 @@
-﻿using escout.Models;
+﻿using escout.Controllers.GenericObjects;
+using escout.Models;
 using escoutTests.Resources;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
